Relecture
Problème d'intégration des JDD. mettre les id au format uuid.
RAF : corriger les classes de la partie service

+++ parent/composant-business/document-interface/src/main/java/cmr/notep/modele/PrecoMouvementsQtes.java
@@ -16,8 +16,8 @@ public class PrecoMouvementsQtes {
     private String id ;
     private int qteMin ;
     private int qteMax ;
-    private int montantMin ;
-    private int montantMax ;
+    private double montantMin ;
+    private double montantMax ;
     private PrecoMouvements precoMouvement ;
     private List<Familles> familles;
     private Ressources ressource;
